test(governance): split API status lifecycle coverage
Move synthetic ProposalApiService status translation scenarios into GovernanceProposalApiStatusIT so API mapping is tested separately from DevKit proposal lifecycle indexing.

Keep GovernanceProposalLifecycleIT focused on real DevKit proposal expiry, status-row stability, duplicate-row protection, and refund behavior. Remove the superseded InfoProposalLifeCycleIT coverage.
